tests: optional execution of external services tests.
* Moves all external services tests to a separate test file.
* Adds script for executing only external services tests.
* Excludes external services tests from main tests script.
* Allows travis failure for the external services tests.
