+++ pytest.ini
@@ -26,7 +26,7 @@
 pep8ignore =
     docs/conf.py ALL
     config.py E501
-addopts = --pep8 --doctest-glob="*.rst" --doctest-modules --cov=rero_ils --cov-report=term-missing --ignore=setup.py
+addopts = --pep8 --doctest-glob="*.rst" --doctest-modules --cov=rero_ils --cov-report=term-missing --ignore=setup.py -m "not external"
 testpaths = docs tests rero_ils
 
 # not displaying all the PendingDeprecationWarnings from invenio
